Move environment utils to separate file
